requirements.txt
requests==2.32.3
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.3.0
markdownify==0.13.1
pandas==2.2.2
python-dateutil==2.9.0.post0